Show UTC format timestamp in diffgraph tooltip, Move getUTCDateFormat to utils
